refactor: refactor onChainReceive screen to the screens folder handle conditional rendering of button center align text and button at button in onchainReceive Screen resolve problem of loading state while creating invoice signed-off-by: Pavan Joshi <pavanj914@gmail.com>
